response status exceptions left

+++ shoppingapp/src/main/java/com/hypersrot/shoppingapp/service/OrderService.java
@@ -5,5 +5,5 @@ import com.hypersrot.shoppingapp.modals.Order;
 public interface OrderService {
     String placeOrder(Integer userId, Integer quantity, String coupon) throws Exception;
 
-    String makePayment(Integer userId, Integer orderId, Integer amount) throws Exception;
+    String makePayment(Integer userId, Integer orderId, Double amount) throws Exception;
 }
